Final Version For User Evaluation 2
Changes made:
Controls:
•	The player’s spaceship now moves at a slightly slower speed, resulting in a more controlled and smoother movement
•	The player's spaceship shoots bullets at a constant rate when simply pressing and holding down the Fire button
•	Shooting lasers can now also be achieved using the Right-mouse click along with the "F" button, allowing users to control the spaceship’s weapons solely via a mouse
Gameplay mechanics:
•	Wormholes are now accompanied with a warning symbol that flashes red and yellow and changes in size, alerting the player of where a wormhole will spawn
•	The overall difficulty of the game has been adjusted by limiting the range of numbers that the random generator can choose from in terms of enemy health points and bullet damage
•	The Informational splash screen now appears for a longer time and also notifies users that they can access it via the main menu screen
•	Asteroids now traverse the screen at a slower speed
•	Power-Ups are now spawned on a reward-basis when the following criteria are fulfilled:
o	Health Regenerating and Bullet Damage Increase power ups are spawned based on number of enemies that are defeated. The random generator determines unique values for each power-up’s target of defeated enemies.
o	Bullet Speed Increase power ups are spawned based on number of asteroids destroyed by the player. The random generator determines the number of asteroids that need to be destroyed before the power-up spawns.
o	Health Increase power-ups are spawned based on number of wormholes destroyed. The random generator determines the number of wormholes that need to be destroyed before the power-up spawns.
o	The special Multiplier Power-Up is spawned based on the sum of wormholes, enemies and asteroids have been defeated. The random generator determines the number of wormholes, asteroids and enemies that need to be destroyed before the power-up spawns.

Code has been adjusted to meet coding standards
Removal of unecessary files